refactor(tests): update logo, toolbar and index tests
